requirements.txt
anyio==3.6.2
fastapi==0.95.1
httpx==0.24.1
pydantic==1.10.7
pymongo==4.3.3
pytest==7.3.1
pytest-asyncio==0.21.0
pytest-mock==3.10.0
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.28.2
urllib3==1.26.15
uvicorn==0.22.0